+++ src/main/java/autopartner/repository/CarModelRepository.java
@@ -3,6 +3,6 @@ package autopartner.repository;
 import autopartner.domain.entity.CarModel;
 import org.springframework.data.repository.CrudRepository;
 
-public interface CarModelRepository extends CrudRepository<CarModel, Integer> {
+public interface CarModelRepository extends CrudRepository<CarModel, Long> {
     Iterable<CarModel> findByActiveTrue();
 }
