+++ src/main/java/autopartner/repository/OrderMaterialRepository.java
@@ -3,6 +3,6 @@ package autopartner.repository;
 import autopartner.domain.entity.OrderMaterial;
 import org.springframework.data.repository.CrudRepository;
 
-public interface OrderMaterialRepository extends CrudRepository<OrderMaterial, Integer> {
+public interface OrderMaterialRepository extends CrudRepository<OrderMaterial, Long> {
     Iterable<OrderMaterial> findByActiveTrue();
 }
