+++ src/main/java/autopartner/service/CarService.java
@@ -6,7 +6,7 @@ public interface CarService {
 
     Iterable<Car> getByActiveTrue();
 
-    Car getCarById(Integer id);
+    Car getCarById(Long id);
 
     Car saveCar(Car car);
